pattern: move Pattern into it's own module
